+++ changelog.md
@@ -1,3 +1,8 @@
+# v0.3
+- Bug squashing
+- Added custom dev console
+- Added more upgrades to idle section
+
 ## v0.2
 - Optimized the shiz outa the physics/rendering engine
 - More balance changes to orb values.